requirements.txt
pandas
numpy
json
re
surprise
surprise.model_selection
collections
sklearn.metrics.pairwise
pickle
ast
sklearn.feature_extraction.text